+++ libcloud/compute/drivers/ec2.py
@@ -3074,7 +3074,8 @@ class BaseEC2NodeDriver(NodeDriver):
         return self._get_terminate_boolean(res)
 
     def create_volume(self, size, name, location=None, snapshot=None,
-                      ex_volume_type='standard', ex_iops=None):
+                      ex_volume_type='standard', ex_iops=None,
+                      ex_encrypted=None, ex_kms_key_id=None):
         """
         Create a new volume.
 
@@ -3104,6 +3105,18 @@ class BaseEC2NodeDriver(NodeDriver):
                      is io1.
         :type iops: ``int``
 
+        :param ex_encrypted: Specifies whether the volume should be encrypted.
+        :type ex_encrypted: ``bool``
+
+        :param ex_kms_key_id: The full ARN of the AWS Key Management
+                            Service (AWS KMS) customer master key (CMK) to use
+                            when creating the encrypted volume.
+                            Example:
+                            arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123
+                            -456a-a12b-a123b4cd56ef.
+                            Only used if encrypted is set to True.
+        :type ex_kms_key_id: ``str``
+
         :return: The newly created volume.
         :rtype: :class:`StorageVolume`
         """
@@ -3128,6 +3141,12 @@ class BaseEC2NodeDriver(NodeDriver):
 
         if ex_volume_type == 'io1' and ex_iops:
             params['Iops'] = ex_iops
+
+        if ex_encrypted is not None:
+            params['Encrypted'] = 1
+
+        if ex_kms_key_id is not None:
+            params['KmsKeyId'] = ex_kms_key_id
 
         volume = self._to_volume(
             self.connection.request(self.path, params=params).object,