+++ libcloud/dns/types.py
@@ -49,7 +49,8 @@ class Provider(object):
     GODADDY = 'godaddy'
     CLOUDFLARE = 'cloudflare'
     NSONE = 'nsone'
-
+    LUADNS = 'luadns'
+    NFSN = 'nfsn'
     # Deprecated
     RACKSPACE_US = 'rackspace_us'
     RACKSPACE_UK = 'rackspace_uk'